requirements.txt
aiofiles==24.1.0
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asyncpg==0.30.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.3
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
kombu==5.4.2
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pytesseract==0.3.13
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.0
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
vine==5.1.0
wcwidth==0.2.13
